requirements.txt
torch 
transformers
plotly
numpy
openpyxl
streamlit 
streamlit-authenticator
streamlit-chat
streamlit_pills
deep_translator
google-cloud-bigquery
db-dtypes
langchain
openai
pandasai
pdfplumber
pypdf
tiktoken
youtube-transcript-api
faiss-cpu
